# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example",
        "description": "Example API Documentation",
        "contact": {
            "name": "Example",
            "email": "example@test.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "api/regions/galleryList": {
            "get": {
                "tags": [
                    "Gallery"
                ],
                "summary": "특정 지역의 갤러리 리스트 조회",
                "operationId": "2754fa0a1f9eb2aa9440009a339a186e",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "조회할 갤러리가 속한 지역의 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공적으로 갤러리 리스트를 가져옴",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "region_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Art Gallery"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "A place for art exhibitions"
                                            },
                                            "manager_id": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "잘못된 요청 파라미터",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "부적절한 접근입니다."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "갤러리가 존재하지 않음",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "갤러리가 존재하지 않습니다."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}